+++ tests/tackle_targets_tests.ads
@@ -1,0 +1,7 @@
+with Testy.Tests;
+
+package Tackle_Targets_Tests is
+   use Testy.Tests;
+
+   procedure Test_Target_Info (T : in out Test_Context);
+end Tackle_Targets_Tests;